+++ src/angular/builder.rs
@@ -18,21 +18,27 @@ pub(crate) use background::stop_background_process;
 
 #[allow(clippy::same_functions_in_if_condition)]
 pub(crate) fn build(
+	#[cfg_attr(not(all(unix, feature = "background")), allow(unused_variables))]
 	ctx: &RenderContext,
 	config: &Config,
 	chapters: Vec<ChapterWithCodeBlocks>,
 ) -> Result<()> {
 	if config.background {
-		if !cfg!(unix) {
-			warn!("The background option is not supported on windows");
-		} else if !cfg!(feature = "background") {
-			warn!("This build doesn't support the background option");
-		} else if !config.experimental_builder {
-			warn!("The background option requires the experimentalBuilder option");
-		} else if ctx.config.get("output.html.live-reload-endpoint").is_none() {
-			warn!("The background option is ignored for commands that don't watch");
-		} else {
-			return self::background::build(config, chapters);
+		#[cfg(not(unix))]
+		warn!("The background option is not supported on windows");
+
+		#[cfg(not(feature = "background"))]
+		warn!("This build doesn't support the background option");
+
+		#[cfg(all(unix, feature = "background"))]
+		{
+			if !config.experimental_builder {
+				warn!("The background option requires the experimentalBuilder option");
+			} else if ctx.config.get("output.html.live-reload-endpoint").is_none() {
+				warn!("The background option is ignored for commands that don't watch");
+			} else {
+				return self::background::build(config, chapters);
+			}
 		}
 	}